格式化代码

+++ src/main/java/com/yqc/forkJoin/CountTask.java
@@ -47,8 +47,8 @@ public class CountTask extends RecursiveTask<Integer> {
             leftTask.fork();
             rightTask.fork();
             // 等待子任务执行完，并得到其结果
-            int leftResult = (int) leftTask.join();
-            int rightResult = (int) rightTask.join();
+            int leftResult = leftTask.join();
+            int rightResult = rightTask.join();
             // 合并子任务
             sum = leftResult + rightResult;
         }
